feat: updated starter screens to use the wdiget
