+++ xact_frontend/lib/screens/start/get_started.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:xact_frontend/screens/startscreen/playnow_screen.dart';
+import 'package:xact_frontend/screens/start/playnow_screen.dart';
+import 'package:xact_frontend/widgets/xact_branding.dart';
 
 class GetStartedScreen extends StatelessWidget {
   const GetStartedScreen({super.key});
@@ -7,41 +8,14 @@ class GetStartedScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: const Color(0xFF1A1F2E),
+      backgroundColor: XActBranding.backgroundColor,
       body: SafeArea(
         child: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 24.0),
           child: Column(
             children: [
               const SizedBox(height: 24),
-              _buildLogo(),
-              const SizedBox(height: 16),
-              const Text(
-                'X-ACT',
-                style: TextStyle(
-                  fontSize: 48,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.white,
-                  letterSpacing: 2,
-                ),
-              ),
-              const SizedBox(height: 8),
-              // Subtitle
-              const Text(
-                'Digital Scotland Yard',
-                style: TextStyle(
-                  fontSize: 18,
-                  color: Colors.white70,
-                  letterSpacing: 1,
-                ),
-              ),
-              const SizedBox(height: 12),
-              // Description
-              const Text(
-                'Hunt down Mister X in this real-world chase game',
-                style: TextStyle(fontSize: 14, color: Colors.white54),
-                textAlign: TextAlign.center,
-              ),
+              XActBranding.buildHeader(),
               const SizedBox(height: 24),
               // Hide Help Button
               _buildHideHelpButton(context),
@@ -53,28 +27,11 @@ class GetStartedScreen extends StatelessWidget {
               _buildPlayNowButton(context),
               const SizedBox(height: 24),
               // Footer text
-              const Text(
-                'Play responsibly • Stay safe • Follow local laws',
-                style: TextStyle(fontSize: 12, color: Colors.white38),
-              ),
+              XActBranding.buildFooter(),
               const SizedBox(height: 16),
             ],
           ),
         ),
-      ),
-    );
-  }
-
-  Widget _buildLogo() {
-    return Container(
-      width: 80,
-      height: 80,
-      decoration: BoxDecoration(
-        shape: BoxShape.circle,
-        border: Border.all(color: const Color(0xFFE53935), width: 3),
-      ),
-      child: const Center(
-        child: Icon(Icons.location_on, color: Color(0xFFE53935), size: 40),
       ),
     );
   }
@@ -89,7 +46,7 @@ class GetStartedScreen extends StatelessWidget {
         },
         style: OutlinedButton.styleFrom(
           side: const BorderSide(color: Colors.white24),
-          backgroundColor: const Color(0xFF252A3A),
+          backgroundColor: XActBranding.cardColor,
           shape: RoundedRectangleBorder(
             borderRadius: BorderRadius.circular(12),
           ),
@@ -108,7 +65,7 @@ class GetStartedScreen extends StatelessWidget {
       width: double.infinity,
       padding: const EdgeInsets.all(20),
       decoration: BoxDecoration(
-        color: const Color(0xFF252A3A),
+        color: XActBranding.cardColor,
         borderRadius: BorderRadius.circular(12),
       ),
       child: Column(
@@ -133,7 +90,7 @@ class GetStartedScreen extends StatelessWidget {
           width: 28,
           height: 28,
           decoration: const BoxDecoration(
-            color: Color(0xFFE53935),
+            color: XActBranding.primaryRed,
             shape: BoxShape.circle,
           ),
           child: Center(
@@ -162,29 +119,14 @@ class GetStartedScreen extends StatelessWidget {
   }
 
   Widget _buildPlayNowButton(BuildContext context) {
-    return SizedBox(
-      width: double.infinity,
-      height: 64,
-      child: ElevatedButton(
-        onPressed: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(builder: (context) => const PlayNowScreen()),
-          );
-        },
-        style: ElevatedButton.styleFrom(
-          backgroundColor: const Color(0xFFE53935),
-          foregroundColor: Colors.white,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(12),
-          ),
-          elevation: 0,
-        ),
-        child: const Text(
-          'Play Now',
-          style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-        ),
-      ),
+    return XActBranding.buildPrimaryButton(
+      text: 'Play Now',
+      onPressed: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(builder: (context) => const PlayNowScreen()),
+        );
+      },
     );
   }
 }
